Threads Reporter added
Thread reporter is added to Shtirlitz llibrary.
This reporter dumps an information about managed threads in a current
proccess. To do this Thread reporter launchs ThreadsWatcher application
(included in Shtirlitz repository) in a separate process.
Important note: to proper execution it is necessary to place following
files in a directory where executable file of analyzed program is
located:
- ThreadsWatcher.exe
- Microsoft.Samples.Debugging.CorApi.dll
- Microsoft.Samples.Debugging.CorApi.dll
- Microsoft.Samples.Debugging.MdbgEngine.dll
- Microsoft.Samples.Debugging.MdbgEngine.xml
- Microsoft.Samples.Debugging.Native.dll